ci: Cleanup broken attempts to test global arrays (ExtConfigKey, ExtConfigDesc)
These did not work due to the execution process context between the tests and the
extension itself being different.

In short: the Mocha test suite code runs in a different process context to the
Extension Host process inside the launched VSCode UI instance under test.

As such, we needed to refactor our global config arrays into a ConfigManager
class, and export the instance of this class from our extension via the
`activate()` function.  Then, the tests can get this instance and use it to
check the contents of a loaded config file.

Finally, a class is a singleton object in TypeScript/JavaScript, which opens up
possibilities to further instrument, mock, or monkey-patch so long as it is done
before the extension is loaded and activated by VSCode's Extension Host process.